move main page components to main folder
