--- src/components/MainPageComponents/InfoPanel/index.ts
@@ -1,1 +1,0 @@
-export { default } from './InfoPanel';
