--- src/components/MainPageComponents/LeaderBoard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './LeaderBoard';
