--- src/components/MainPageComponents/MainMenuItem/index.ts
@@ -1,1 +1,0 @@
-export { default } from './MainMenuItem';
