+++ src/components/index.ts
@@ -5,4 +5,5 @@ export * from './auth';
 export * from './common';
 export * from './forum';
 export * from './game';
+export * from './main';
 export * from './profile';
