+++ src/components/main/InfoPanel/InfoPanel.css
