+++ src/components/main/InfoPanel/index.ts
@@ -1,0 +1,1 @@
+export * from './InfoPanel';
