+++ src/components/main/LeaderBoard/index.ts
@@ -1,0 +1,1 @@
+export * from './LeaderBoard';
