+++ src/components/main/MainMenuItem/index.ts
@@ -1,0 +1,1 @@
+export * from './MainMenuItem';
