+++ src/components/main/index.ts
@@ -1,0 +1,3 @@
+export * from './MainMenuItem';
+export * from './LeaderBoard';
+export * from './InfoPanel';
